+++ api_service/templates/base.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8"/>
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <meta http-equiv="Content-Security-Policy" content="upgrade-insecure-requests">
-    <link href="{{ url_for('static', path='/cash_receipt_editor.css') }}" rel="stylesheet"/>
+    <link href="{{ url_for('static', path='/css/cash_receipt_editor.css') }}" rel="stylesheet"/>
     <link id="favicon" rel="icon" type="image/x-icon" href="{{ url_for('static', path='/favicon4.ico') }}"/>
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <title>ЦифроХаб Service</title>
