+++ api_service/templates/cash_receipt_editor.html
