+++ static/css/cash_receipt_editor.css
