--- static/styles.css
@@ -1,172 +1,0 @@
-body {
-    background: #e2eaf0;
-}
-
-li {
-	list-style-type: none;
-  }
-
-.navbar {
-    background-color: #00111a;
-}
-
-.nav-link {
-	padding: 1px;
-	color: #ffffff;
-}
-
-.nav-link.nav-link-yellow {
-	padding: 1px;
-	color: #f6ff00;
-}
-
-.nav-link.nav-link-white {
-	padding: 1px;
-	color: #ffffff;
-}
-
-.btn {
-	width: 300px;
-	font-family: Roboto;
-	font-size: 16;
-}
-
-.container-fluid {
-	justify-content: center;
-	padding-left: 0px;
-}
-
-.container.container-banner {
-	max-width: 898px;
-}
-
-.container .main-photo {
-	box-shadow: 5px 5px 5px rgb(101, 112, 128);
-	max-width: 898px;
-}
-
-.carousel-inner{
-	box-shadow: 5px 5px 5px rgb(101, 112, 128);
-}
-
-header .navbar {
-    font-style: normal;
-    font-weight: normal;
-    font-size: 20px;
-    line-height: 20px;
-    padding: 0px 0rem;
-	margin-bottom: 10px;
-
-}
-
-.adress {
-	color: #f0f4f7;
-	padding-left: 10px;
-}
-
-main {
-	max-width: 898px;
-}
-
-.container-fluid {
-	min-width: 320px;
-    display: flex;
-}  
-
-.main-contain {
-	margin-top: 10px;
-	margin-bottom: 10px;
-	display: flex;
-	flex-wrap: wrap;
-	justify-content: center;
-}
-
-.stock-contain {
-	display: flex;
-	flex-wrap: wrap;
-	justify-content: center;
-}
-
-.card {
-	width: 27rem;
-	border: none;
-}
-
-.btn-circle:hover,.btn-circle:focus,.btn-circle:visited {
-	background-color: rgb(115, 138, 148);
-	border-color: #ffffff;
-}
-
-.btn-circle {
-    width: 125px;
-    height: 70px;
-    border-radius: 5px;
-	border-color: rgb(101, 112, 128);
-    text-align: center;
-	margin: 4px;
-    padding-left: 0;
-    padding-right: 0;
-    font-size: 13px;
-	color: #ffffff;
-	background-color: rgb(98, 111, 129);
-	box-shadow: 2px 2px 4px rgb(70, 83, 141);
-}
-
-.card-text {
-	height: 48px;
-	color: #3B4A59;
-	font-size: 19px;
-	text-align: center;
-	line-height: 99%;
-}
-
-.card-quantity {
-	font-family: Roboto;
-	font-size: 14px;
-	text-align: center;
-}
-
-.text-body-secondary {
-	font-size: 14px;
-}
-
-.card-price {
-	color: #12171b;
-	text-align: center;
-	font-size: 25px;
-	background-color: #e5e9ec;
-
-}
-
-.card-description {
-	font-family: Roboto;
-	font-size: 14px;
-}
-
-.hover-image-scale {
-	width: 100%;
-	height: 100%;
-	display: inline-block; 
-	position: relative;
-}
-  
-.stock-contain img {
-	transition: 0.4s;
-	display: block; 
-}
-
-.stock-contain img:hover {
-	border-radius: 20%;
-	transform: scale(1.12); /* Увеличиваем масштаб */
-}
-
-.btn-circle:hover {
-	transform: scale(1.02); /* Увеличиваем масштаб */
-	transition: 0.1s;
-}
-
-.contacts {
-	text-align: center;
-	color: black;
-	padding: 12px 12px;
-}
