--- templates/index.html
@@ -1,97 +1,0 @@
-<!DOCTYPE html>
-<html lang="ru">
-
-<head>
-  <meta charset="UTF-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <meta http-equiv="Content-Security-Policy" content="upgrade-insecure-requests">
-  <meta name="description" content="Ленино Крым продажа смартфонов настройка аксессуары" />
-  <link href="{{ url_for('static', path='/bootstrap.min.css') }}" rel="stylesheet" />
-  <link href="{{ url_for('static', path='/styles.css') }}" rel="stylesheet" />
-  <link id="favicon" rel="icon" type="image/x-icon" href="{{ url_for('static', path='/favicon.ico') }}" />
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <title>Салон Цифротех</title>
-</head>
-
-<body>
-  <header>
-    <nav class="navbar navbar-expand-sm">
-      <div class="container">
-        <div class="collapse navbar-collapse" id="navbarSupportedContent">
-          <div class="row">
-            <div class="col-sm-12 d-flex justify-content-end">
-              <a class="navbar-brand" href="/address/">
-                <img class="logo-anim" src="{{ url_for('static', path='/logo.png') }}" width="200" height="44"
-                  loading="lazy">
-                <div class="adress">
-                  <small>
-                    <h6>п. Ленино, пр-т Ленина 9</h6>
-                  </small>
-                </div>
-              </a>
-            </div>
-          </div>
-        </div>
-        <div class="container-fluid">
-          <ul class="navbar-nav mr-auto mb-2 mb-lg-0">
-            <li class="nav-item">
-              <a class="nav-link nav-link-yellow" aria-current="page" href="/">
-                <h6>КАТАЛОГ ТОВАРОВ</h6>
-              </a>
-            </li>
-            <li class="nav-item">
-              <a class="nav-link nav-link-white" aria-current="page" href="/12">
-                <h6>СМАРТФОНЫ</h6>
-              </a>
-
-            <li class="nav-item">
-              <a class="nav-link nav-link-yellow" href="/36">
-                <h6>УМНЫЕ ЧАСЫ</h6>
-              </a>
-            </li>
-            <li class="nav-item">
-              <a class="nav-link nav-link-white" href="/29">
-                <h6>ПЛАНШЕТЫ</h6>
-              </a>
-            </li>
-            <li class="nav-item">
-              <a class="nav-link nav-link-yellow" href="/28">
-                <h6>КНОПОЧНЫЕ ТЕЛЕФОНЫ</h6>
-              </a>
-            </li>
-          </ul>
-        </div>
-      </div>
-
-    </nav>
-  </header>
-  <div class="container container-banner">
-    <div id="carouselExampleIndicators" class="carousel slide carousel-fade" data-bs-ride="carousel">
-      <div class="carousel-inner">
-        <div class="carousel-item active">
-          <img src="{{ url_for('static', path='/slide1.jpg') }}" class="d-block w-100" alt="slide1" />
-        </div>
-        <div class="carousel-item">
-          <img src="{{ url_for('static', path='/slide2.jpg') }}" class="d-block w-100" alt="slide2" />
-        </div>
-        <div class="carousel-item">
-          <img src="{{ url_for('static', path='/slide3.jpg') }}" class="d-block w-100" alt="slide3" />
-        </div>
-        <button class="carousel-control-prev" type="button" data-bs-target="#carouselExampleIndicators"
-          data-bs-slide="prev">
-          <span class="carousel-control-prev-icon" aria-hidden="false"></span>
-          <span class="visually-hidden">Previous</span>
-        </button>
-        <button class="carousel-control-next" type="button" data-bs-target="#carouselExampleIndicators"
-          data-bs-slide="next">
-          <span class="carousel-control-next-icon" aria-hidden="false"></span>
-          <span class="visually-hidden">Next</span>
-        </button>
-      </div>
-      <main>{% block content %} {% endblock %}</main>
-    </div>
-  </div>
-  <script src="{{ url_for('static', path='/bootstrap.min.js') }}"></script>
-</body>
-
-</html>
