--- templates/product.html
@@ -1,31 +1,0 @@
-{% extends 'index.html' %}
-{% block content %}
-
-
-<div class="main-contain">
-    {% for button in menu_buttons %}
-    <a href="/{{ button.code }}">
-        <button class="btn btn-dark btn-circle" class="btn btn-primary btn-circle">{{ button.name }}</button>
-    </a>
-    {% endfor %}
-</div>
-
-<div class="stock-contain">
-    {% for var in data %}
-    <div class="card">
-        <img src="photo/{{ var.code }}.jpg" class="card-img-top" alt="{{ var.name }}">
-        <div class="card-body">
-            <div class="card-text">{{ var.name }}</div>
-            <div class="card-price">{{ var.price }} р. </div>
-            <div class="card-quantity"><small class="text-body-secondary">В наличии {{ var.quantity
-                    }}
-                    шт.</small></div>
-            {% if var.desc %}
-            <div class="card-description">{{ var.desc[0] |safe }}</div>
-            {% endif %}
-        </div>
-    </div>
-    {% endfor %}
-</div>
-
-{% endblock %}
