--- templates/products_list.html
@@ -1,8 +1,0 @@
-{% extends 'index.html' %}
-{% block content %}
-{% for button in buttons %}
-<div class="container container-banner">
-    <a class="btn btn-light" href="/{{button.code }}" role="button">{{ button.name }}</a>
-</div>
-{% endfor %}
-{% endblock %}
